fix: clean up plugin loading and fix VSCode extension (#312)
* fix: clean up plugin loading and fix VSCode extension

* address pr comments
