+++ packages/cli/src/actions/action-utils.ts
@@ -1,13 +1,10 @@
-import { createZModelServices, loadDocument, type ZModelServices } from '@zenstackhq/language';
-import { isDataSource, isPlugin, Model } from '@zenstackhq/language/ast';
-import { getLiteral } from '@zenstackhq/language/utils';
+import { loadDocument } from '@zenstackhq/language';
+import { isDataSource } from '@zenstackhq/language/ast';
 import { PrismaSchemaGenerator } from '@zenstackhq/sdk';
 import colors from 'colors';
 import fs from 'node:fs';
 import path from 'node:path';
-import { fileURLToPath } from 'node:url';
 import { CliError } from '../cli-error';
-import { PLUGIN_MODULE_NAME } from '../constants';
 
 export function getSchemaFile(file?: string) {
     if (file) {
@@ -37,9 +34,7 @@ export function getSchemaFile(file?: string) {
 }
 
 export async function loadSchemaDocument(schemaFile: string) {
-    const { ZModelLanguage: services } = createZModelServices();
-    const pluginDocs = await getPluginDocuments(services, schemaFile);
-    const loadResult = await loadDocument(schemaFile, pluginDocs);
+    const loadResult = await loadDocument(schemaFile);
     if (!loadResult.success) {
         loadResult.errors.forEach((err) => {
             console.error(colors.red(err));
@@ -50,63 +45,6 @@ export async function loadSchemaDocument(schemaFile: string) {
         console.warn(colors.yellow(warn));
     });
     return loadResult.model;
-}
-
-export async function getPluginDocuments(services: ZModelServices, fileName: string): Promise<string[]> {
-    // parse the user document (without validation)
-    const parseResult = services.parser.LangiumParser.parse(fs.readFileSync(fileName, { encoding: 'utf-8' }));
-    const parsed = parseResult.value as Model;
-
-    // balk if there are syntax errors
-    if (parseResult.lexerErrors.length > 0 || parseResult.parserErrors.length > 0) {
-        return [];
-    }
-
-    // traverse plugins and collect "plugin.zmodel" documents
-    const result: string[] = [];
-    for (const decl of parsed.declarations.filter(isPlugin)) {
-        const providerField = decl.fields.find((f) => f.name === 'provider');
-        if (!providerField) {
-            continue;
-        }
-
-        const provider = getLiteral<string>(providerField.value);
-        if (!provider) {
-            continue;
-        }
-
-        let pluginModelFile: string | undefined;
-
-        // first try to treat provider as a path
-        let providerPath = path.resolve(path.dirname(fileName), provider);
-        if (fs.existsSync(providerPath)) {
-            if (fs.statSync(providerPath).isDirectory()) {
-                providerPath = path.join(providerPath, 'index.js');
-            }
-
-            // try plugin.zmodel next to the provider file
-            pluginModelFile = path.resolve(path.dirname(providerPath), PLUGIN_MODULE_NAME);
-            if (!fs.existsSync(pluginModelFile)) {
-                // try to find upwards
-                pluginModelFile = findUp([PLUGIN_MODULE_NAME], path.dirname(providerPath));
-            }
-        }
-
-        if (!pluginModelFile) {
-            // try loading it as a ESM module
-            try {
-                const resolvedUrl = import.meta.resolve(`${provider}/${PLUGIN_MODULE_NAME}`);
-                pluginModelFile = fileURLToPath(resolvedUrl);
-            } catch {
-                // noop
-            }
-        }
-
-        if (pluginModelFile && fs.existsSync(pluginModelFile)) {
-            result.push(pluginModelFile);
-        }
-    }
-    return result;
 }
 
 export function handleSubProcessError(err: unknown) {
