+++ packages/language/src/zmodel-workspace-manager.ts
@@ -1,7 +1,6 @@
 import {
     DefaultWorkspaceManager,
     URI,
-    UriUtils,
     type AstNode,
     type LangiumDocument,
     type LangiumDocumentFactory,
@@ -11,9 +10,7 @@ import type { LangiumSharedServices } from 'langium/lsp';
 import fs from 'node:fs';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { isPlugin, type Model } from './ast';
-import { PLUGIN_MODULE_NAME, STD_LIB_MODULE_NAME } from './constants';
-import { getLiteral } from './utils';
+import { STD_LIB_MODULE_NAME } from './constants';
 
 export class ZModelWorkspaceManager extends DefaultWorkspaceManager {
     private documentFactory: LangiumDocumentFactory;
@@ -71,87 +68,5 @@ export class ZModelWorkspaceManager extends DefaultWorkspaceManager {
 
         const stdlib = await this.documentFactory.fromUri(URI.file(stdLibPath));
         collector(stdlib);
-
-        const documents = this.langiumDocuments.all;
-        const pluginModels = new Set<string>();
-
-        // find plugin models
-        documents.forEach((doc) => {
-            const parsed = doc.parseResult.value as Model;
-            parsed.declarations.forEach((decl) => {
-                if (isPlugin(decl)) {
-                    const providerField = decl.fields.find((f) => f.name === 'provider');
-                    if (providerField) {
-                        const provider = getLiteral<string>(providerField.value);
-                        if (provider) {
-                            pluginModels.add(provider);
-                        }
-                    }
-                }
-            });
-        });
-
-        if (pluginModels.size > 0) {
-            console.log(`Used plugin modules: ${Array.from(pluginModels)}`);
-
-            // the loaded plugin models would be removed from the set
-            const pendingPluginModules = new Set(pluginModels);
-
-            await Promise.all(
-                folders
-                    .map((wf) => [wf, this.getRootFolder(wf)] as [WorkspaceFolder, URI])
-                    .map(async (entry) => this.loadPluginModels(...entry, pendingPluginModules, collector)),
-            );
-        }
-    }
-
-    protected async loadPluginModels(
-        workspaceFolder: WorkspaceFolder,
-        folderPath: URI,
-        pendingPluginModels: Set<string>,
-        collector: (document: LangiumDocument) => void,
-    ): Promise<void> {
-        const content = (await this.fileSystemProvider.readDirectory(folderPath)).sort((a, b) => {
-            // make sure the node_modules folder is always the first one to be checked
-            // so we can exit early if the plugin is found
-            if (a.isDirectory && b.isDirectory) {
-                const aName = UriUtils.basename(a.uri);
-                if (aName === 'node_modules') {
-                    return -1;
-                } else {
-                    return 1;
-                }
-            } else {
-                return 0;
-            }
-        });
-
-        for (const entry of content) {
-            if (entry.isDirectory) {
-                const name = UriUtils.basename(entry.uri);
-                if (name === 'node_modules') {
-                    for (const plugin of Array.from(pendingPluginModels)) {
-                        const path = UriUtils.joinPath(entry.uri, plugin, PLUGIN_MODULE_NAME);
-                        try {
-                            await this.fileSystemProvider.readFile(path);
-                            const document = await this.langiumDocuments.getOrCreateDocument(path);
-                            collector(document);
-                            console.log(`Adding plugin document from ${path.path}`);
-
-                            pendingPluginModels.delete(plugin);
-                            // early exit if all plugins are loaded
-                            if (pendingPluginModels.size === 0) {
-                                return;
-                            }
-                        } catch {
-                            // no-op. The module might be found in another node_modules folder
-                            // will show the warning message eventually if not found
-                        }
-                    }
-                } else {
-                    await this.loadPluginModels(workspaceFolder, entry.uri, pendingPluginModels, collector);
-                }
-            }
-        }
     }
 }
